+++ src/oci/database/models/autonomous_database.py
@@ -474,6 +474,10 @@ class AutonomousDatabase(object):
             The value to assign to the connection_urls property of this AutonomousDatabase.
         :type connection_urls: oci.database.models.AutonomousDatabaseConnectionUrls
 
+        :param public_connection_urls:
+            The value to assign to the public_connection_urls property of this AutonomousDatabase.
+        :type public_connection_urls: oci.database.models.AutonomousDatabaseConnectionUrls
+
         :param license_model:
             The value to assign to the license_model property of this AutonomousDatabase.
             Allowed values for this property are: "LICENSE_INCLUDED", "BRING_YOUR_OWN_LICENSE", 'UNKNOWN_ENUM_VALUE'.
@@ -503,6 +507,10 @@ class AutonomousDatabase(object):
         :param private_endpoint:
             The value to assign to the private_endpoint property of this AutonomousDatabase.
         :type private_endpoint: str
+
+        :param public_endpoint:
+            The value to assign to the public_endpoint property of this AutonomousDatabase.
+        :type public_endpoint: str
 
         :param private_endpoint_label:
             The value to assign to the private_endpoint_label property of this AutonomousDatabase.
@@ -835,6 +843,7 @@ class AutonomousDatabase(object):
             'service_console_url': 'str',
             'connection_strings': 'AutonomousDatabaseConnectionStrings',
             'connection_urls': 'AutonomousDatabaseConnectionUrls',
+            'public_connection_urls': 'AutonomousDatabaseConnectionUrls',
             'license_model': 'str',
             'used_data_storage_size_in_tbs': 'int',
             'freeform_tags': 'dict(str, str)',
@@ -842,6 +851,7 @@ class AutonomousDatabase(object):
             'subnet_id': 'str',
             'nsg_ids': 'list[str]',
             'private_endpoint': 'str',
+            'public_endpoint': 'str',
             'private_endpoint_label': 'str',
             'private_endpoint_ip': 'str',
             'db_version': 'str',
@@ -951,6 +961,7 @@ class AutonomousDatabase(object):
             'service_console_url': 'serviceConsoleUrl',
             'connection_strings': 'connectionStrings',
             'connection_urls': 'connectionUrls',
+            'public_connection_urls': 'publicConnectionUrls',
             'license_model': 'licenseModel',
             'used_data_storage_size_in_tbs': 'usedDataStorageSizeInTBs',
             'freeform_tags': 'freeformTags',
@@ -958,6 +969,7 @@ class AutonomousDatabase(object):
             'subnet_id': 'subnetId',
             'nsg_ids': 'nsgIds',
             'private_endpoint': 'privateEndpoint',
+            'public_endpoint': 'publicEndpoint',
             'private_endpoint_label': 'privateEndpointLabel',
             'private_endpoint_ip': 'privateEndpointIp',
             'db_version': 'dbVersion',
@@ -1066,6 +1078,7 @@ class AutonomousDatabase(object):
         self._service_console_url = None
         self._connection_strings = None
         self._connection_urls = None
+        self._public_connection_urls = None
         self._license_model = None
         self._used_data_storage_size_in_tbs = None
         self._freeform_tags = None
@@ -1073,6 +1086,7 @@ class AutonomousDatabase(object):
         self._subnet_id = None
         self._nsg_ids = None
         self._private_endpoint = None
+        self._public_endpoint = None
         self._private_endpoint_label = None
         self._private_endpoint_ip = None
         self._db_version = None
@@ -2200,6 +2214,30 @@ class AutonomousDatabase(object):
         self._connection_urls = connection_urls
 
     @property
+    def public_connection_urls(self):
+        """
+        Gets the public_connection_urls of this AutonomousDatabase.
+        The Public URLs of Private Endpoint database for accessing Oracle Application Express (APEX) and SQL Developer Web with a browser from a Compute instance within your VCN or that has a direct connection to your VCN.
+
+
+        :return: The public_connection_urls of this AutonomousDatabase.
+        :rtype: oci.database.models.AutonomousDatabaseConnectionUrls
+        """
+        return self._public_connection_urls
+
+    @public_connection_urls.setter
+    def public_connection_urls(self, public_connection_urls):
+        """
+        Sets the public_connection_urls of this AutonomousDatabase.
+        The Public URLs of Private Endpoint database for accessing Oracle Application Express (APEX) and SQL Developer Web with a browser from a Compute instance within your VCN or that has a direct connection to your VCN.
+
+
+        :param public_connection_urls: The public_connection_urls of this AutonomousDatabase.
+        :type: oci.database.models.AutonomousDatabaseConnectionUrls
+        """
+        self._public_connection_urls = public_connection_urls
+
+    @property
     def license_model(self):
         """
         Gets the license_model of this AutonomousDatabase.
@@ -2436,6 +2474,30 @@ class AutonomousDatabase(object):
         :type: str
         """
         self._private_endpoint = private_endpoint
+
+    @property
+    def public_endpoint(self):
+        """
+        Gets the public_endpoint of this AutonomousDatabase.
+        The public endpoint for the private endpoint enabled resource.
+
+
+        :return: The public_endpoint of this AutonomousDatabase.
+        :rtype: str
+        """
+        return self._public_endpoint
+
+    @public_endpoint.setter
+    def public_endpoint(self, public_endpoint):
+        """
+        Sets the public_endpoint of this AutonomousDatabase.
+        The public endpoint for the private endpoint enabled resource.
+
+
+        :param public_endpoint: The public_endpoint of this AutonomousDatabase.
+        :type: str
+        """
+        self._public_endpoint = public_endpoint
 
     @property
     def private_endpoint_label(self):